+++ js/jquery.terminal-1.9.0.js
@@ -32,7 +32,7 @@
  * Copyright (c) 2007-2013 Alexandru Marasteanu <hello at alexei dot ro>
  * licensed under 3 clause BSD license
  *
- * Date: Sat, 21 Oct 2017 07:44:00 +0000
+ * Date: Sat, 21 Oct 2017 14:16:35 +0000
  */
 
 /* TODO:
@@ -841,11 +841,11 @@
     // alert only first exception of type
     // ---------------------------------------------------------------------
     var excepctions = [];
-    function alert_exception(e, label) {
+    function alert_exception(label, e) {
         var message = (label ? label + ': ' : '') + exception_message(e);
         if (excepctions.indexOf(message) === -1) {
             excepctions.push(message);
-            alert(message + '\n' + e.stack);
+            alert(message + (e.stack ? '\n' + e.stack : ''));
         }
     }
     // ---------------------------------------------------------------------
@@ -4495,7 +4495,11 @@
             var options = {
                 finalize: function finalize(div) {
                     a11y_hide(div.addClass('command'));
-                    settings.onEchoCommand.call(self, div);
+                    try {
+                        settings.onEchoCommand.call(self, div, command);
+                    } catch(e) {
+                        console.log(e.stack);
+                    }
                 }
             };
             if ($.isFunction(prompt)) {